requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
grad-cam==1.4.8
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.0
lightning-utilities==0.9.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
ttach==0.0.3
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.212
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
